# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task API
  description: >
    A small, multi-user HTTP REST API for task management, built with Go
    using only the standard library plus pgx (PostgreSQL) and
    golang.org/x/crypto (bcrypt). Every task belongs to exactly one user
    and is only ever visible to that user. Each task has a title, an
    optional description, a status (`pending`, `in_progress`, `done`, or
    `cancelled`, following the transition rules below), and a priority
    (`low`, `medium`, or `high`). Authentication is opaque bearer session
    tokens, issued by POST /auth/login.


    This document describes exactly the behavior implemented in
    `internal/task/handler.go`+`service.go`+`repository.go` and
    `internal/user/handler.go`+`service.go`+`repository.go` as of the
    current codebase — no aspirational or planned behavior is included.


    Every response from every endpoint — including error responses and
    responses produced before routing, such as a `404` for an unknown path
    — additionally carries a fixed set of security headers
    (`middleware.SecurityHeaders`): `Content-Security-Policy: default-src
    'none'; frame-ancestors 'none'`, `X-Frame-Options: DENY`, and
    `X-Content-Type-Options: nosniff`. They are constant, identical
    everywhere, and are not part of the API contract a client programs
    against, so they are documented here once rather than repeated on each
    response. `Strict-Transport-Security` joins them only when the
...
servers:
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Create a user account
      security: []
      description: >
        Creates a new user account. Does not log the caller in — call
        POST /auth/login afterwards to obtain a session token.
      requestBody:
        required: true
        description: >
          Request bodies larger than 1 MiB are rejected, surfacing as the
          same 400 "invalid request body" response as malformed JSON —
          the same cap the task endpoints apply.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: alice@example.com
              password: correct horse battery staple
      responses:
        '201':
          description: Account created.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 9f8e7d6c-5b4a-3210-9876-543210fedcba
                email: alice@example.com
                created_at: "2026-08-08T12:00:00Z"
                updated_at: "2026-08-08T12:00:00Z"
        '400':
          description: >
            Malformed JSON body; an empty/malformed `email` (must contain
            "@", at most 320 characters); or a `password` shorter than 8
            or longer than 72 characters (bcrypt's own input limit).
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "invalid input: password must be at least 8 characters"
        '409':
          description: A user with this email is already registered.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: email already registered
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in and obtain a session token
      security: []
      description: >
        Verifies email/password and issues a new bearer session token,
        valid for AUTH_SESSION_TTL (default 24h). The raw token is
        returned exactly once, in this response — it cannot be recovered
        later; only its hash is stored server-side. Send it back as
        `Authorization: Bearer <token>` on every subsequent request to a
        protected endpoint.
      requestBody:
        required: true
        description: >
          Request bodies larger than 1 MiB are rejected, surfacing as the
          same 400 "invalid request body" response as malformed JSON.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: alice@example.com
              password: correct horse battery staple
      responses:
        '200':
          description: Authenticated successfully.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                token: 0094162c1a4c9c2e5aa5897e55ea477e6dd089a60e2d060303c82ff1fe82e6d
                expires_at: "2026-08-09T12:00:00Z"
                user:
                  id: 9f8e7d6c-5b4a-3210-9876-543210fedcba
                  email: alice@example.com
                  created_at: "2026-08-08T12:00:00Z"
                  updated_at: "2026-08-08T12:00:00Z"
        '400':
          description: Malformed JSON body.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid request body
        '401':
          description: >
            Unknown email or wrong password — deliberately the same
            response either way, so a caller can never distinguish "no
            such account" from "wrong password" and enumerate registered
            emails.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: invalid email or password
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Invalidate the current session token
      description: >
        Deletes the session backing the bearer token used for this
        request.

        Repeating it with the *same* token returns 401, not 204: the
        endpoint requires authentication, and the first call already
        destroyed the session that authenticates it. The idempotency here
        is server-side — deleting an already-absent session is not treated
        as an error internally — but it is not observable through this
        endpoint, because a request carrying an unknown or expired token
        is rejected before it reaches the handler. Clients should treat
        both 204 and 401 as "the session is gone", which is the outcome
        logout is asking for either way.
      responses:
        '204':
          description: Logged out. No response body.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get the authenticated user
      responses:
        '200':
          description: The user identified by the bearer token used for this request.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: >
            The token authenticated successfully but its owning user no
            longer exists. Only reachable if the account was removed out
            of band (direct SQL, `cmd/seed -reset`) between the session
            lookup and this read — this API exposes no endpoint that
            deletes a user, and deleting one normally cascades its
            sessions away, which would produce a 401 instead.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: user not found
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Opaque session token returned by POST /auth/login (not a JWT — it
        has no independently verifiable structure; the server looks it up
        by its hash on every request). Send as `Authorization: Bearer
        <token>`.
...
  headers:
    XRequestID:
      description: >
        Correlates this response with server-side structured logs
        (`request_id` field). Echoes the client-supplied `X-Request-Id`
        request header when it is present and well-formed (non-empty,
        at most 128 characters, and restricted to `[A-Za-z0-9_-]`);
        otherwise the server generates a new one. Present on every
        response from every endpoint, including error responses.
      schema:
        type: string
      example: 7a947bb6436a2e5009efcaa55c42cdf8
  schemas:
...
    User:
      type: object
      description: >
        The persisted representation of a user account. Never includes a
        password or password hash — user.User.PasswordHash is
        `json:"-"` and is never serialized to any response.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: 9f8e7d6c-5b4a-3210-9876-543210fedcba
        email:
          type: string
          format: email
          example: alice@example.com
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: "2026-08-08T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          readOnly: true
          example: "2026-08-08T12:00:00Z"
      required:
        - id
        - email
        - created_at
        - updated_at
    RegisterRequest:
      type: object
      description: Accepted body for POST /auth/register.
      properties:
        email:
          type: string
          format: email
          maxLength: 320
          description: Required. Must contain "@". Trimmed of leading/trailing whitespace.
          example: alice@example.com
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: >
            Required, 8–72 characters. The 72-character ceiling is
            bcrypt's own input limit (golang.org/x/crypto/bcrypt rejects
            longer input outright), not an arbitrary policy choice.
          example: correct horse battery staple
      required:
        - email
        - password
    LoginRequest:
      type: object
      description: Accepted body for POST /auth/login.
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          example: correct horse battery staple
      required:
        - email
        - password
    LoginResponse:
      type: object
      description: Body returned by a successful POST /auth/login.
      properties:
        token:
          type: string
          description: >
            Bearer session token. Returned exactly once — only its hash
            is stored server-side, so it cannot be retrieved again after
            this response.
          example: 0094162c1a4c9c2e5aa5897e55ea477e6dd089a60e2d060303c82ff1fe82e6d
        expires_at:
          type: string
          format: date-time
          description: When `token` stops being valid (see AUTH_SESSION_TTL).
          example: "2026-08-09T12:00:00Z"
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - expires_at
        - user
...
    ErrorResponse:
      type: object
      description: >
        The single JSON envelope used by every error response from every
        endpoint in this API.
      properties:
        error:
          type: string
          description: Human-readable error message. Never contains internal details.
          example: task not found
      required:
        - error
  responses:
    Unauthorized:
      description: >
        Missing, malformed, or invalid/expired `Authorization: Bearer`
        header.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: invalid or expired session token
    TooManyRequests:
      description: >
        The client's IP has made too many requests to /auth/register or
        /auth/login (combined — they share one rate-limit budget) within
        the current window (see cmd/api/main.go's
        authRateLimitRequests/authRateLimitWindow). Not returned by any
        other endpoint. Wait and retry.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: too many requests, please try again later
...
    InternalServerError:
      description: >
        An unexpected server-side failure. The response body is always
        this exact generic message — the underlying error is never
        exposed to the client, only logged server-side (with the same
        X-Request-Id, for correlation).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: internal server error
    ServiceUnavailable:
      description: >
        The session lookup that authenticates this request could not be
        performed — the database is unreachable, timed out, or the request
        was canceled. **This says nothing about the caller's token:** a
        perfectly valid, unexpired token gets this response too, so the
        client must NOT discard its credentials and re-authenticate. Retry
        the same request with the same token once the service recovers
        (GET /health/ready reports when that is).

        Returned only by endpoints that require authentication, since it
        originates in the session lookup itself. POST /auth/register and
        POST /auth/login perform no such lookup, so the same underlying
        database failure surfaces there as a 500 instead.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: service temporarily unavailable, please retry